Integrate single file and multi-file generators into the same flow in GenTaskGraph

+++ framework/src/ppl/delite/framework/codegen/delite/generators/DeliteGenTaskGraph.scala
@@ -157,14 +157,12 @@ trait DeliteGenTaskGraph extends DeliteCodegen with DeliteKernelCodegen with Loo
         case _ => Config.buildDir + sep + gen + sep + "kernels" + sep
       }
       val outDir = new File(buildPath); outDir.mkdirs()
-      val outFile = new File(buildPath + kernelName + "." + gen.fileExtension)
-      val kStream = new PrintWriter(outFile)
+      val kernelFile = buildPath + kernelName + "." + gen.fileExtension
+      val (outFile, kStream) = gen.getFileStream(kernelFile)
       try {
         gen.withStream(kStream){ gen.emitKernel(sym, rhs) }
-        kStream.close()
-
-        if (gen.emitSingleFile) {
-          outFile.delete()
+        if (!gen.emitSingleFile) {
+          kStream.close()
         }
 
         // record that this kernel was successfully generated
@@ -179,8 +177,10 @@ trait DeliteGenTaskGraph extends DeliteCodegen with DeliteKernelCodegen with Loo
 
       } catch {
         case e: GenerationFailedException => // no generator found
-          kStream.close()
-          outFile.delete()
+          if (!gen.emitSingleFile) {
+            kStream.close()
+            outFile.delete()
+          }
           if (Config.dumpException) {
             System.err.println(gen.toString + ":" + (sym.map(quote)))
             e.printStackTrace
